refactor: extract handler request-building, restore dedup naming, move SLA breach domain types
